rename ABI files to match versions and consolidate mapping files
