+++ packages/subgraph/networks.yaml
@@ -6,25 +6,16 @@ mainnet:
     bondingManager:
       address: '511bc4556d823ae99630ae8de28b9b80df90ea2e'
       startBlock: 5533958
-    bondingManager_LIP11:
-      address: '511bc4556d823ae99630ae8de28b9b80df90ea2e'
-      startBlock: 5533958
-    bondingManager_streamflow:
-      address: '511bc4556d823ae99630ae8de28b9b80df90ea2e'
-      startBlock: 5533958
     roundsManager:
-      address: '3984fc4ceeef1739135476f625d36d6c35c40dc3'
-      startBlock: 5533997
-    roundsManager_streamflow:
       address: '3984fc4ceeef1739135476f625d36d6c35c40dc3'
       startBlock: 5533997
     livepeerToken:
       address: '58b6a8a3302369daec383334672404ee733ab239'
       startBlock: 5533907
-    minter:
+    minterV1:
       address: '8573f2f5a3bd960eee3d998473e50c75cdbe6828'
       startBlock: 5533913
-    minter_LIP40:
+    minter:
       address: '505F8c2ee81f1C6fa0D88e918eF0491222E05818'
       startBlock: 10686186
     ticketBroker:
@@ -45,25 +36,16 @@ rinkeby:
     bondingManager:
       address: 'e75a5DccfFe8939F7f16CC7f63EB252bB542FE95'
       startBlock: 5325862
-    bondingManager_LIP11:
-      address: 'e75a5DccfFe8939F7f16CC7f63EB252bB542FE95'
-      startBlock: 5325862
-    bondingManager_streamflow:
-      address: 'e75a5DccfFe8939F7f16CC7f63EB252bB542FE95'
-      startBlock: 5325862
     roundsManager:
-      address: '572d1591bD41f50130FD0212058eAe34F1B17290'
-      startBlock: 5325866
-    roundsManager_streamflow:
       address: '572d1591bD41f50130FD0212058eAe34F1B17290'
       startBlock: 5325866
     livepeerToken:
       address: '23b814a57D53b1a7A860194F53401D0D639abED7'
       startBlock: 5325849
-    minter:
+    minterV1:
       address: 'e5056d1fE48F562dd8AcB1BBFFa147223d448C1B'
       startBlock: 5325851
-    minter_LIP40:
+    minter:
       address: '63B12E71F9c620BeAdf53b0e95189a16A2Af8Ed2'
       startBlock: 7041202
     ticketBroker:
@@ -84,25 +66,16 @@ development:
     bondingManager:
       address: 'A94B7f0465E98609391C623d0560C5720a3f2D33'
       startBlock: 0
-    bondingManager_LIP11:
-      address: 'A94B7f0465E98609391C623d0560C5720a3f2D33'
-      startBlock: 0
-    bondingManager_streamflow:
-      address: 'A94B7f0465E98609391C623d0560C5720a3f2D33'
-      startBlock: 0
     roundsManager:
-      address: '5f8e26fAcC23FA4cbd87b8d9Dbbd33D5047abDE1'
-      startBlock: 0
-    roundsManager_streamflow:
       address: '5f8e26fAcC23FA4cbd87b8d9Dbbd33D5047abDE1'
       startBlock: 0
     livepeerToken:
       address: 'D833215cBcc3f914bD1C9ece3EE7BF8B14f841bb'
       startBlock: 0
-    minter:
+    minterV1:
       address: 'e982E462b094850F12AF94d21D470e21bE9D0E9C'
       startBlock: 0
-    minter_LIP40:
+    minter:
       address: 'e982E462b094850F12AF94d21D470e21bE9D0E9C'
       startBlock: 0
     ticketBroker:
